move WarehouseDetails from pages to components
