+++ src/components/WarehouseDetails/WarehouseDetails.scss
